requirements.txt
aiohttp==3.8.5
pytest==3.2.1
pytest-asyncio==0.6.0
pydest
python-dotenv
flask
